Change behavior to send cert via event bus
This set of changes works to move responsibility of writing
certificates from the Salt Reactor system to the requesting
application. To accomplish this, the response containing the
signed certificate is sent back via the Event bus to the minion
requesting a certificate.

This is a hack, make no doubt, but there is not a clean way to
call a salt module like this via Python. Or maybe there is and
I was not able to find documentation. The road I was going down
was to assemble a salt.util.event.MasterEvent payload, but I could
not get the minion to receieve events. Only with this call via
the CLI, so we're spawning to do the needful